refactor(daemon): extract MaintenancePipelineRunner into scheduler/ (batch B)
Second batch of the AceClawDaemon decomposition. Lifts the
runLearningMaintenancePipeline orchestration (150 LoC) +
triggerMaintenanceDraftLifecycle helper (32 LoC) + the inner
LearningMaintenanceRunBuilder class (85 LoC) + the pure summarizeTrends
helper (5 LoC) into a dedicated runner under a new
dev.aceclaw.daemon.scheduler subpackage.

What moved:
- runLearningMaintenancePipeline (150 LoC) -> MaintenancePipelineRunner.run
- triggerMaintenanceDraftLifecycle (32 LoC) -> private helper
- LearningMaintenanceRunBuilder inner class (85 LoC) -> private static
  inner class
- summarizeTrends pure helper (5 LoC) -> package-private static

What stayed:
- Lambda call site in wireAgentHandler now reads:
  `(trigger, scope) -> maintenanceRunner.run(trigger, scope.workspaceHash(), scope.workingDir())`
  -- the runner is constructed once at wireup, deps captured via the new
  Deps record.

Dependencies model:
The pipeline pulls in 15 collaborators. Inlining a Deps record on the
runner (rather than a 15-arg constructor) keeps the call-site readable
and makes the contract self-documenting via per-field javadocs noting
which collaborators are required vs nullable-as-off-switch.

Behavior preservation:
- All 6 sub-stages keep their try/catch isolation (one stage failure
  doesn't abort the rest -- background maintenance must be tolerant).
- All nullability conventions preserved verbatim
  (historicalLogIndex/candidateStore/journal/etc).
- The bug fix from #513 (auto-promotion workspace attribution) does NOT
  apply here -- the maintenance pipeline only knows one workspace per
  trigger fire, so workingDir is the correct parameter throughout.

New tests:
4 unit tests for summarizeTrends in MaintenancePipelineRunnerTest:
- empty list -> empty string
- under 3 items: joins all
- over 3 items: truncates to first 3
- order preserved
The pipeline's stage orchestration is covered by the existing daemon
end-to-end integration tests (they exercise the scheduler tick).

LoC:
- AceClawDaemon: 2,909 -> 2,638 (-271, -9.3%)
- Cumulative (batches A + B): 3,112 -> 2,638 (-474, -15.2%)
- New MaintenancePipelineRunner: ~360 LoC (heavy javadoc on Deps record)
- New MaintenancePipelineRunnerTest: ~50 LoC
